Day 03, 2016 Challenge Solution (#5)
This one was a bit tricky. It took me quite a while to debug my triangles.valid() method. I was making a mistake with isosceles triangles by assuming they would always be valid when of course they would not be. The trick with the triangles is that there can be multiple max length sides.